Test coverage for beat clock

+++ odin-core/src/main/java/com/purplepip/odin/sequence/measure/StaticBeatMeasureProvider.java
@@ -28,7 +28,7 @@ public class StaticBeatMeasureProvider implements MeasureProvider {
 
   @Override
   public double getMeasure(double count) {
-    return (long) count / beatsPerMeasure;
+    return count / beatsPerMeasure;
   }
 
   /**
